requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
pypdf==4.0.1
img2pdf==0.5.1
pikepdf==8.11.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
structlog==24.1.0
slowapi==0.1.9
prometheus-client==0.19.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
